software/testproject/code/virtual_language_learning_buddy/virtual_language_learning_buddy/cultural_context.py
```python
class CulturalContext:
    def __init__(self, user_profile, cultural_data):
        """
        Initialize the CulturalContext with user profile and cultural data.

        :param user_profile: Instance of UserProfile containing user preferences.
        :param cultural_data: Dictionary containing cultural context data.
        """
        self.user_profile = user_profile
        self.cultural_data = cultural_data

    def provide_contextual_response(self, input_text):
        """
        Provide a response based on the cultural context and user input.

        :param input_text: The input text from the user.
        :return: A culturally relevant response.
        """
        language = self.user_profile.language
        context_preference = self.user_profile.context_preference

        # Retrieve cultural context based on user preferences
        context = self.cultural_data.get(language, {}).get(context_preference, "general")

        # Generate a response based on the cultural context
        response = f"In the context of {context}, here's a response to your input: {input_text}"

        return response
```

### Explanation

- **CulturalContext Class**: This class is responsible for providing culturally relevant responses based on the user's profile and cultural data.
  
- **Attributes**:
  - `user_profile`: An instance of `UserProfile` that contains user preferences such as language and context preference.
  - `cultural_data`: A dictionary containing cultural context data, which is used to tailor responses.

- **Methods**:
  - `__init__`: Initializes the class with a user profile and cultural data.
  - `provide_contextual_response`: Takes user input text and provides a response based on the cultural context. It retrieves the appropriate cultural context using the user's language and context preference, then generates a response.

This implementation assumes that `cultural_data` is a nested dictionary where the first level keys are languages, and the second level keys are context preferences, each mapping to a specific cultural context.